Added functionality to Mission, Criteria, and Ranking pages
Changed the mission page to not save the mission, but instead store it in session storage.
Created the ability to add project criteria and edit the titles. Also created the ability to remove criteria. Planning to add more editing capability (e.g. changing the description).
Created a ranking page where the use will choose which criteria are more important than others.